refactor: Rename directory to router.
